This is a partial implementation of a new feature. The idea is that on create/update of a mailbox, you can mark it as 'deleteConfirmationRequired.' Then, when you go to delete it, you'll first have to confirm by clicking a link that will take you to some lambda that clears the hash in that Dynamo item. I have not yet implemented the lambda that lets you clear that value, nor have I exposed a way to even know what that link is (maybe an email? maybe just a response the first time you try to delete?) For now any items with this column set to some value will fail deletion - as designed.

+++ serverless-step-functions-java/src/main/java/com/godaddy/DeletionHandler.java
@@ -60,7 +60,9 @@ public class DeletionHandler implements RequestHandler<Map<String, Object>, ApiG
 		}
 		
 		// Delete the mailbox
-		deleteItem(id);
+		if (id != null) {
+			deleteItem(id);
+		}
 		
 
 		Map<String, String> headers = new HashMap<>();
